pyproject.toml
[project]
name = "our-a2a-project"
version = "0.1.0"
description = "Claude 기반 A2A 멀티에이전트 시스템"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # A2A SDK 코어 + 서버 의존성 분리 설치
    "a2a-sdk>=0.3.0",
    "starlette>=0.40.0",
    "sse-starlette>=2.0.0",
    # Claude API
    "anthropic>=0.40.0",
    # 서버/유틸리티
    "uvicorn>=0.34.0",
    "click>=8.1.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    # Supabase (Baby Brain DB)
    "supabase>=2.0.0",
    # OpenAI (Embeddings)
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "hosts", "common"]
